pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paint-by-number"
version = "0.0.0"
description = "A paint-by-number image quantization tool."
authors = [{name = "Kasper van Maasdam"}]
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
dependencies = [
    "pillow>=12.0.0,<13.0.0",
    "numpy>=2.4.0,<3.0.0",
    "scikit-image>=0.26.0,<1.0.0",
    "scikit-learn>=1.8.0,<2.0.0",
    "scipy>=1.16.3,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.19.0,<2.0.0"
]

[project.scripts]
pbn = "pbn.cli:main"

[tool.setuptools]
packages = ["pbn"]

[tool.mypy]
ignore_missing_imports = true
warn_unused_configs = true
strict = true
check_untyped_defs = true

disallow_untyped_calls = false
disallow_untyped_defs = false
warn_unused_ignores = true
warn_redundant_casts = true
